B/feat/image

+++ backend/books/views.py
@@ -21,6 +21,8 @@ from .deepL_translation import translate_summary
 
 
 class BookListAPIView(APIView):
+    permission_classes = [IsAuthenticated]
+
     # 전체 목록 조회
     def get(self, request):
         books = Book.objects.order_by("-created_at")
@@ -52,9 +54,16 @@ class BookListAPIView(APIView):
 
 
 class DALL_EImageAPIView(APIView):
+    permission_classes = [IsAuthenticated]
+
     def post(self, request, book_id):
         client = OpenAI()
         book = get_object_or_404(Book, id=book_id)
+        if book.user_id is not request.user:
+            return Response(
+                {"error": "You don't have permission."},
+                status=status.HTTP_401_UNAUTHORIZED,
+            )
         response = client.images.generate(
             model="dall-e-3",
             prompt=f"{book.title}, {book.tone}",
@@ -69,35 +78,38 @@ class DALL_EImageAPIView(APIView):
         serializer = BookSerializer(instance=book)
         return Response(serializer.data, status = 200)
 
+
 class BookDetailAPIView(APIView):
+    permission_classes = [IsAuthenticated]
+
     # 상세 조회
     def get(self, request, book_id):
         book = get_object_or_404(Book, id=book_id)
         ratings = Rating.objects.filter(book=book)
 
         # chapter 내용들 가져오기
-        chapters = Chapter.objects.filter(book_id = book_id)
-        chapter_serializer = ChapterSerializer(chapters, many = True)
+        chapters = Chapter.objects.filter(book_id=book_id)
+        chapter_serializer = ChapterSerializer(chapters, many=True)
 
         # 책 전체 내용 직렬화
         book_serializer = BookSerializer(book)
 
         response_data = book_serializer.data
-        response_data['chapters'] = chapter_serializer.data
+        response_data["chapters"] = chapter_serializer.data
         return Response(response_data, status=200)
 
     # chapter(summary) 생성
     def post(self, request, book_id):
-        summary = request.data.get("summary")
-        language = request.data.get("language", "EN-US")
-        if not summary:
+        book = get_object_or_404(Book, id=book_id)
+        if book.user_id != request.user:
             return Response(
-                {"error": "Missing summary prompt"}, status=status.HTTP_400_BAD_REQUEST
+                {"error": "You don't have permission."},
+                status=status.HTTP_401_UNAUTHORIZED,
             )
 
+        language = request.data.get("language", "EN-US")
         chapter = Chapter.objects.filter(book_id=book_id).last()
         if not chapter:
-            book = get_object_or_404(Book, id=book_id)
             elements = ElementsSerializer(book)
             chapter_num = 0
             result = prologue_generator(elements.data)
@@ -111,17 +123,21 @@ class BookDetailAPIView(APIView):
                 )
             chapter_num = chapter.chapter_num
             result = summary_generator(chapter_num, summary)
+            chapter_num += 1
             content = result["final_summary"]
 
-        translated_content = translate_summary(content,language)
+        translated_content = translate_summary(content, language)
         if not translated_content:
             return Response(
                 {"error": "Translation failed or empty result"},
-                status=status.HTTP_500_INTERNAL_SERVER_ERROR
+                status=status.HTTP_500_INTERNAL_SERVER_ERROR,
             )
 
         serializer = ChapterSerializer(
-            data={"content": translated_content, "book_id": book_id, "chapter_num" : chapter_num}
+            data={
+                "content": translated_content,
+                "book_id": book_id,
+            }
         )
         if serializer.is_valid(raise_exception=True):
             serializer.save()
@@ -133,8 +149,13 @@ class BookDetailAPIView(APIView):
     # 글 수정
     def put(self, request, book_id):
         book = get_object_or_404(Book, id=book_id)
-        serializer = BookSerializer(book, data=request.data, partial=True)
+        if book.user_id != request.user:
+            return Response(
+                {"error": "You don't have permission."},
+                status=status.HTTP_401_UNAUTHORIZED,
+            )
 
+        serializer = BookSerializer(book, data=request.data, partial=True)
         if serializer.is_valid(raise_exception=True):
             serializer.save()
             return Response(serializer.data, status=201)
@@ -143,38 +164,25 @@ class BookDetailAPIView(APIView):
     # 글 삭제
     def delete(self, request, book_id):
         book = get_object_or_404(Book, id=book_id)
+        if book.user_id != request.user:
+            return Response(
+                {"error": "You don't have permission."},
+                status=status.HTTP_401_UNAUTHORIZED,
+            )
         book.delete()
         return Response("No Content", status=204)
 
-class DeletePrologueAPIView(APIView) :
-    def delete(self, request, book_id) :
-        prologue = Chapter.objects.filter(chapter_num=0, book_id = book_id)
+
+class DeletePrologueAPIView(APIView):
+    def delete(self, request, book_id):
+        prologue = Chapter.objects.filter(chapter_num=0, book_id=book_id)
         prologue.delete()
         return Response("Prologue deleted successfully", status=204)
-        
+
 
 class BookLikeAPIView(APIView):
     permission_classes = [IsAuthenticated]
 
-    def post(self, request, book_id):
-        book = get_object_or_404(Book, id=book_id)
-        like_bool = False
-        # 좋아요 삭제
-        if request.user in book.is_liked.all():
-            book.is_liked.remove(request.user)
-        # 좋아요 추가
-        else:
-            book.is_liked.add(request.user)  
-            like_bool = True
-        serializer = BookLikeSerializer(book)
-        return Response(
-            {
-                "like_bool": like_bool,
-                "total_likes": book.total_likes(),
-                "book": serializer.data,
-            },
-            status=200,
-        )
     def get(self, request, book_id):
         book = get_object_or_404(Book, id=book_id)
         like_bool = request.user in book.is_liked.all()
@@ -188,6 +196,32 @@ class BookLikeAPIView(APIView):
             status=200,
         )
 
+    def post(self, request, book_id):
+        book = get_object_or_404(Book, id=book_id)
+        if book.user_id != request.user:
+            return Response(
+                {"error": "You don't have permission."},
+                status=status.HTTP_401_UNAUTHORIZED,
+            )
+        like_bool = False
+        # 좋아요 삭제
+        if request.user in book.is_liked.all():
+            book.is_liked.remove(request.user)
+        # 좋아요 추가
+        else:
+            book.is_liked.add(request.user)
+            like_bool = True
+        serializer = BookLikeSerializer(book)
+        return Response(
+            {
+                "like_bool": like_bool,
+                "total_likes": book.total_likes(),
+                "book": serializer.data,
+            },
+            status=200,
+        )
+
+
 class UserLikedBooksAPIView(APIView):
     permission_classes = [IsAuthenticated]
 
@@ -199,14 +233,18 @@ class UserLikedBooksAPIView(APIView):
         serializer = BookSerializer(book_likes, many=True)
         return Response(serializer.data, status=200)
 
+
 class UserBooksAPIView(APIView):
     permission_classes = [IsAuthenticated]
 
     def get(self, request):
         user = request.user
-        user_books = user.books.all()  # 역참조를 이용해 사용자가 작성한 책 리스트를 가져옴
+        user_books = (
+            user.books.all()
+        )  # 역참조를 이용해 사용자가 작성한 책 리스트를 가져옴
         serializer = BookSerializer(user_books, many=True)
         return Response(serializer.data, status=200)
+
 
 class RatingAPIView(APIView):
     permission_classes = [IsAuthenticated]
@@ -221,6 +259,11 @@ class RatingAPIView(APIView):
 
     def post(self, request, book_id):
         book = get_object_or_404(Book, id=book_id)
+        if book.user_id != request.user:
+            return Response(
+                {"error": "You don't have permission."},
+                status=status.HTTP_401_UNAUTHORIZED,
+            )
         rating = request.data.get("rating")
 
         if rating not in [1, 2, 3, 4, 5]:
@@ -240,6 +283,8 @@ class RatingAPIView(APIView):
 
 
 class CommentListAPIView(APIView):
+    permission_classes = [IsAuthenticated]
+
     def get(self, request, book_id):
         book = get_object_or_404(Book, id=book_id)
         comments = book.comments.all()
@@ -255,8 +300,15 @@ class CommentListAPIView(APIView):
 
 
 class CommentDetailAPIView(APIView):
+    permission_classes = [IsAuthenticated]
+
     def put(self, request, book_id, comment_id):
         comment = get_object_or_404(Comment, id=comment_id)
+        if comment.user_id != request.user:
+            return Response(
+                {"error": "You don't have permission."},
+                status=status.HTTP_401_UNAUTHORIZED,
+            )
         serializer = CommentSerializer(comment, data=request.data, partial=True)
         if serializer.is_valid(raise_exception=True):
             serializer.save()
@@ -264,5 +316,10 @@ class CommentDetailAPIView(APIView):
 
     def delete(self, request, book_id, comment_id):
         comment = get_object_or_404(Comment, id=comment_id)
+        if comment.user_id != request.user:
+            return Response(
+                {"error": "You don't have permission."},
+                status=status.HTTP_401_UNAUTHORIZED,
+            )
         comment.delete()
         return Response("NO comment", status=204)
